Another try to fix MSVC compilation

+++ Tests/src/file_exchange/test_file_send_receive_utils.cpp
@@ -287,7 +287,7 @@ static FileExchangeTestResult runFileExchangeTest(ClientStorage& clientStorage, 
 					return;
 				}
 
-				fileToWriteIdx = it->second;
+				fileToWriteIdx = static_cast<int>(it->second);
 				fileCursor = 0;
 			},
 			.getFileLength = [&filesToSend, &fileToWriteIdx](std::ifstream&) -> uint64_t {
